Isoler la logique physique dans un module dédié
